requirements.txt
fastapi
uvicorn
python-dotenv
torch
transformers
accelerate
 language-tool-python
